feat: main function implementation
