+++ Sources/Amplitude/Events/RevenueEvent.swift
@@ -1,0 +1,18 @@
+//
+//  RevenueEvent.swift
+//
+//
+//  Created by Marvin Liu on 11/3/22.
+//
+
+import Foundation
+
+public class RevenueEvent: BaseEvent {
+    override public var eventType: String {
+        get {
+            return "revenue_amount"
+        }
+        set {
+        }
+    }
+}
